Initial merge of antirez\3.0
Conflicts:
	src/anet.c
	src/aof.c
	src/bitops.c
	src/config.c
	src/db.c
	src/debug.c
	src/dict.c
	src/migrate.c
	src/object.c
	src/redis.c
	src/redis.h
	src/sentinel.c
	src/t_list.c
	src/t_set.c
	src/t_zset.c
	src/util.c
	tests/instances.tcl

+++ tests/instances.tcl
@@ -102,7 +102,13 @@ proc spawn_instance {type base_port count {conf {}}} {
 proc cleanup {} {
     puts "Cleaning up..."
     foreach pid $::pids {
-        catch {exec kill -9 $pid}
+        catch {
+			if { $::tcl_platform(platform) == "windows" } {
+				kill_proc2 $pid
+			} else {
+				exec kill -9 $pid
+			}
+		}
     }
     foreach dir $::dirs {
         catch {exec rm -rf $dir}
@@ -395,8 +401,11 @@ proc kill_instance {type id} {
     if {$pid == -1} {
         error "You tried to kill $type $id twice."
     }
-
-    exec kill -9 $pid
+    if { $::tcl_platform(platform) == "windows" } {
+        kill_proc2 $pid
+    } else {
+        exec kill -9 $pid
+    }
     set_instance_attrib $type $id pid -1
     set_instance_attrib $type $id link you_tried_to_talk_with_killed_instance
